Prevent non-error warnings with Bluejax.
Bluejax will warn if a Promise is rejected with a non-error. It has its own
notion of that an "Error" is. `x instanceof Error` is not enough. There has to
be a `message`, `stack` and a `name` (among other things). Our problem was that
on some platforms `stack` was not populated. On other platforms, `name` was not
populated either. This commit fixes the problems.